extractscripts/requirements.txt
google-generativeai
python-dotenv
PyPDF2
python-docx
pathlib
flask
flask-cors
fastapi
uvicorn[standard]
python-multipart
pydantic